♻️ refactor(eval): move agent-eval-run routes into the hono workflows router
The 10 `agent-eval-run` endpoints were Next.js route segments using
`@upstash/workflow/nextjs`, while every other workflow already lives in the
`/api/workflows` hono app. Move them to
`apps/server/src/router-hono/workflows/agent-eval-run` and let the existing
`[[...route]]` catch-all serve them.

- `serve()` switches from `@upstash/workflow/nextjs` to `@upstash/workflow/hono`
- the two completion webhooks (`on-*-complete`) become plain hono handlers
- `qstashClient` from `@/libs/qstash` is kept as-is: unlike
  `createWorkflowQstashClient()`, it carries the Vercel protection-bypass header
  these workflows rely on
- URLs are unchanged, so `AgentEvalRunWorkflow.trigger*` callers stay put

Adds a route-mount regression test pinning the hono mount against
`WORKFLOW_PATHS`, since a path typo now 404s at runtime instead of failing the
build.
