+++ apps/server/src/router-hono/workflows/index.ts
@@ -1,5 +1,6 @@
 import { Hono } from 'hono';
 
+import agentEvalRunApp from './agent-eval-run';
 import agentSignalApp from './agent-signal';
 import expertiseHistoryApp from './expertise-history';
 import memoryUserMemoryApp from './memory-user-memory';
@@ -11,6 +12,7 @@ import verifyApp from './verify';
 
 const app = new Hono().basePath('/api/workflows');
 
+app.route('/agent-eval-run', agentEvalRunApp);
 app.route('/agent-signal', agentSignalApp);
 app.route('/expertise-history', expertiseHistoryApp);
 app.route('/memory-user-memory', memoryUserMemoryApp);
